🔧 Fix compilation errors
- Fixed format string in rust_dirac_delta.rs
- Fixed import paths for LatticePointDerive
- Removed problematic Eq/Hash traits from structs with f32/f64 fields
- Changed f32/f64 to u32 for hashable structs
- Fixed string comparison in eigenmatrix projection

All core modules now compile cleanly.

+++ compiler/introspector-collector/src/github_ecosystem_lattice.rs
@@ -1,4 +1,4 @@
-use crate::lattice_macros::LatticePointDerive;
+use crate::meta_lattice::LatticePointDerive;
 
 #[derive(Debug, Clone, PartialEq, Eq, Hash, LatticePointDerive)]
 pub enum GitHubEcosystem {
@@ -61,12 +61,12 @@ pub struct PullRequestNode {
     pub state: String,
 }
 
-#[derive(Debug, Clone, PartialEq, Eq, Hash, LatticePointDerive)]
+#[derive(Debug, Clone, PartialEq)]
 pub struct LanguageStats {
     pub repo: String,
     pub language: String,
     pub bytes: u64,
-    pub percentage: f32,
+    pub percentage: u32, // Changed from f32 to u32
 }
 
 #[derive(Debug, Clone, PartialEq, Eq, Hash, LatticePointDerive)]
